Fix calculation of ISO-TP SF size field length
Given the following setup:
* CAN-FD bus
* tx_data_length = 64
* extended addressing

When sending a 7 byte payload, the full ISO-TP frame with size and address fields
would have a minimum size of 7 + 1 (addr) + 1 (size) = 9. According to the standard,
such a payload must use a two-byte size field. However, the current implementation
does not correctly calculate the length of the size field.

This commit fixes the calculation and introduces a unit test to verify the correct
behavior.